requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
Markdown==3.2.1
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.1
uWSGI==2.0.18
whitenoise==3.3.1